sync icon, level, and name and rank >= 30
